Add files via upload
Comprehensive zip file of the submission as downloaded from the Udacity Workspaces. User can also find random images from the internet used in the notebook for test purpose; as well as a JSON file that maps category names of the training dataset to the actual flowers names. The later was provided by Udacity as part of the project instructions.